backend-old-basic/requirements.txt
# Conecta Plus API - Dependencies

# =============================================================================
# Core Framework
# =============================================================================
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==23.0.0
pydantic==2.9.0
pydantic-settings==2.5.0

# =============================================================================
# Database
# =============================================================================
sqlalchemy[asyncio]==2.0.32
asyncpg==0.29.0
alembic==1.13.2
psycopg2-binary==2.9.9  # Para Alembic sync operations

# =============================================================================
# Cache & Queue
# =============================================================================
redis==5.1.0
aioredis==2.0.1

# =============================================================================
# Auth & Security
# =============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Pinned for passlib compatibility
python-multipart==0.0.9

# =============================================================================
# HTTP Client
# =============================================================================
httpx==0.27.2
aiohttp==3.9.5

# =============================================================================
# Utils
# =============================================================================
python-dateutil==2.9.0
python-dotenv==1.0.1

# =============================================================================
# Excel/CSV/Reports
# =============================================================================
openpyxl==3.1.5

# =============================================================================
# Logging
# =============================================================================
structlog==24.4.0

# =============================================================================
# Rate Limiting
# =============================================================================
slowapi==0.1.9

# =============================================================================
# Monitoring & Health
# =============================================================================
prometheus-fastapi-instrumentator==7.0.0

# =============================================================================
# Testing
# =============================================================================
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2  # Also used for TestClient
factory-boy==3.3.1
faker==30.3.0

# =============================================================================
# Development
# =============================================================================
black==24.8.0
isort==5.13.2
flake8==7.1.1
mypy==1.11.2

# Stress Testing
locust>=2.20.0
email-validator>=2.0.0
